+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.21.0
 ### Features
 * HOLD keymap modifier + HOLD+[SHIFT]+BACKSPACE/DELETE to delete word before and after the cursor [#420](https://github.com/jcubic/jquery.terminal/issues/420)
+* align tabs like in unix terminal [#423](https://github.com/jcubic/jquery.terminal/issues/423)
 
 ### Bugfix
 * fix &) in scheme prism formatting [#421](https://github.com/jcubic/jquery.terminal/issues/421)